# swagger.json
{
	"swagger": "2.0",
	"info": {
		"title": "Recipe Book API",
		"description": "Recipe Book API",
		"version": "1.0.0"
	},
	"host": "recipe-book-00.herokuapp.com",
	"basePath": "/",
	"schemes": ["https"],
	"securityDefinitions": {
		"oAuthSample": {
			"type": "oauth2",
			"authorizationUrl": "https://recipe-book-00.herokuapp.com/api-docs",
			"flow": "implicit",
			"scopes": {
				"read_recipes": "read your recipes",
				"write_recipes": "modify recipes in your account"
			}
		},
		"api_key": {
			"type": "apiKey",
			"name": "api_key",
			"in": "header",
			"api_key": "thisisanapikey123"
		}
	},
	"security": [
		{
			"oAuthSample": ["write_recipes", "read_recipes"]
		}
	],
	"paths": {
		"/": {
			"get": {
				"description": "",
				"parameters": [],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Must be logged in to access"
					}
				}
			}
		},
		"/recipes/": {
			"get": {
				"description": "",
				"parameters": [],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Must be logged in to access"
					}
				}
			},
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"schema": {
							"type": "object",
							"properties": {
								"name": {
									"example": "any"
								},
								"ingredients": {
									"example": "any"
								},
								"steps": {
									"example": "any"
								}
							}
						}
					}
				],
				"responses": {
					"201": {
						"description": "Created"
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Must be logged in to access"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/recipes/{id}": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Must be logged in to access"
					}
				}
			},
			"put": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"name": "body",
						"in": "body",
						"schema": {
							"type": "object",
							"properties": {
								"name": {
									"example": "any"
								},
								"ingredients": {
									"example": "any"
								},
								"steps": {
									"example": "any"
								}
							}
						}
					}
				],
				"responses": {
					"204": {
						"description": "No Content"
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Must be logged in to access"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			},
			"delete": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"204": {
						"description": "No Content"
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Must be logged in to access"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		}
	}
}